LayoutTests/webaudio/mediaelementaudiosourcenode-expected.txt

Basic tests for MediaElementAudioSourceNode API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Source AudioNode has no inputs.
PASS Source AudioNode has one output.
PASS connect() exception thrown for illegal destination AudioNode.
PASS connect() exception thrown for illegal output index.
PASS connect() exception thrown for illegal input index.
PASS audioNode.connect(context.destination) succeeded.
PASS createMediaElementSource() threw error when called twice on same HTMLMediaElement.
PASS context.createMediaElementSource(null) threw exception TypeError: Argument 1 ('mediaElement') to AudioContext.createMediaElementSource must be an instance of HTMLMediaElement.
PASS successfullyParsed is true

TEST COMPLETE
